Tool, ToolCallStatus, ToolCall and AssistantResult now live in one module
