+++ server/api/views/assistant/tool_services.py
@@ -1,66 +1,11 @@
-from dataclasses import dataclass
-from typing import Callable
-
-# search_documents is defined in search_tool.py; this `from ... import` binds a local
-# name, api.views.assistant.tool_services.search_documents, that SEARCH_TOOL.run looks
-# up at call time. Tests patch that local name — the use site here, NOT the definition
-# site (search_tool.search_documents) — because mock.patch must rebind the reference the
-# code actually resolves. Keep this as a bare-name import: rewriting SEARCH_TOOL.run to
-# call search_tool.search_documents(...) would move the patch target and break the tests.
+from api.views.assistant.assistant_types import Tool
+# Keep this as a bare-name import: rewriting SEARCH_TOOL.run to call
+# search_tool.search_documents(...) would move the patch target and break the tests.
 from api.views.assistant.search_tool import search_documents
 # Reuse the existing ask_database implementation from services/tools rather than
 # reimplementing it here — it already enforces the SELECT-only and ALLOWED_TABLES
-# guards, and does no DB work at import time. No test patches this one today, but the
-# same use-site rule would apply if one did: the patch target is
-# api.views.assistant.tool_services.ask_database (the name bound here), not the
-# definition in api.services.tools.database. Keep it a bare-name import so that
-# stays true.
+# guards, and does no DB work at import time.
 from api.services.tools.database import ask_database
-
-
-@dataclass(frozen=True)
-class Tool:
-    """One assistant tool: the schema the model sees (data) and the function we run
-    (behavior), bundled together under a single name.
-
-    Bundling name/description/parameters/run in one object means each tool is
-    registered in exactly one place — the TOOLS list at the bottom of this module —
-    so the schema sent to the model and the callable actually invoked can never drift
-    apart. Adding a tool is appending one Tool to TOOLS; nothing else changes.
-
-    Behavior is stored as the `run` field (composition) rather than a method on a
-    subclass because our tools differ only in *which* function runs — same schema()
-    machinery, same fields, just a different callable. They are instances of one
-    concept, not distinct kinds of thing.
-
-    TODO: Flip to `Tool(ABC)` + one subclass per tool (with `run` as a method) if a
-    tool ever needs more than a swapped-in function — specifically when it:
-      - carries per-type state/setup (a client, connection, cache, validated config);
-      - overrides more than run (e.g. a custom schema() shape, or extra methods like
-        validate_arguments / cost_estimate);
-      - needs a per-type run signature or an @abstractmethod-enforced contract so a
-        tool with no behavior fails at class-definition time, not at call time.
-    Until then the callable field is lighter and keeps registration drift-proof.
-    """
-
-    name: str
-    description: str
-    parameters: dict
-    # run(user, **arguments) -> str. Every tool takes the request `user` so the dispatch
-    # loop can call them uniformly; a tool that doesn't need it simply ignores it.
-    run: Callable
-
-    def schema(self) -> dict:
-        # Flattened Responses-API shape: name/description/parameters at the top level.
-        # This is intentionally NOT the nested {"function": {...}} shape that the Chat
-        # Completions API (and services/tools/tools.py's create_tool_dict) uses.
-        return {
-            "type": "function",
-            "name": self.name,
-            "description": self.description,
-            "parameters": self.parameters,
-        }
-
 
 
 SEARCH_TOOL = Tool(
@@ -104,7 +49,7 @@ Be specific rather than generic - use terms that would appear in the relevant do
 # update above, cheap for a table this small and stable.
 _MEDICATION_SCHEMA_STRING = "Table: api_medication\nColumns: name, benefits, risks"
 
-# TODO: sharpen this description — it is the highest-value change on this branch, and
+# Sharpen this description — it is the highest-value change on this branch, and
 # there are now two eval runs of evidence behind it. Across 14 tool calls over those two
 # runs the model selected ask_database exactly 0 times. Decisively, for "What medications
 # are recommended for bipolar depression?" it generated a semantic search query rather
